Move scaffolding generator code into generator pkg
