+++ pkg/generators/scaffolding/render.go
@@ -1,7 +1,6 @@
 package scaffolding
 
 import (
-	"bytes"
 	"strings"
 	"text/template"
 
@@ -69,24 +68,4 @@ func GraphqlFields(
 	}
 
 	return result
-}
-
-func renderTemplate(
-	name string,
-	input string,
-	data interface{},
-) (*string, error) {
-	t, err := template.New(name).Funcs(funcMap).Parse(input)
-	if err != nil {
-		return nil, err
-	}
-
-	var content bytes.Buffer
-	err = t.Execute(&content, data)
-	if err != nil {
-		return nil, err
-	}
-
-	contStr := content.String()
-	return &contStr, nil
 }
